// src/generated-api/models/CommonResponseScaleQuestionFindAllResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CareDiary API
 * 돌봄일기 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ScaleQuestionFindAllResponse } from './ScaleQuestionFindAllResponse';
import {
    ScaleQuestionFindAllResponseFromJSON,
    ScaleQuestionFindAllResponseFromJSONTyped,
    ScaleQuestionFindAllResponseToJSON,
    ScaleQuestionFindAllResponseToJSONTyped,
} from './ScaleQuestionFindAllResponse';

/**
 * 
 * @export
 * @interface CommonResponseScaleQuestionFindAllResponse
 */
export interface CommonResponseScaleQuestionFindAllResponse {
    /**
     * 
     * @type {number}
     * @memberof CommonResponseScaleQuestionFindAllResponse
     */
    status: number;
    /**
     * 
     * @type {string}
     * @memberof CommonResponseScaleQuestionFindAllResponse
     */
    code: string;
    /**
     * 
     * @type {string}
     * @memberof CommonResponseScaleQuestionFindAllResponse
     */
    message: string;
    /**
     * 
     * @type {ScaleQuestionFindAllResponse}
     * @memberof CommonResponseScaleQuestionFindAllResponse
     */
    data?: ScaleQuestionFindAllResponse;
}

/**
 * Check if a given object implements the CommonResponseScaleQuestionFindAllResponse interface.
 */
export function instanceOfCommonResponseScaleQuestionFindAllResponse(value: object): value is CommonResponseScaleQuestionFindAllResponse {
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('message' in value) || value['message'] === undefined) return false;
    return true;
}

export function CommonResponseScaleQuestionFindAllResponseFromJSON(json: any): CommonResponseScaleQuestionFindAllResponse {
    return CommonResponseScaleQuestionFindAllResponseFromJSONTyped(json, false);
}

export function CommonResponseScaleQuestionFindAllResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommonResponseScaleQuestionFindAllResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'],
        'code': json['code'],
        'message': json['message'],
        'data': json['data'] == null ? undefined : ScaleQuestionFindAllResponseFromJSON(json['data']),
    };
}

export function CommonResponseScaleQuestionFindAllResponseToJSON(json: any): CommonResponseScaleQuestionFindAllResponse {
    return CommonResponseScaleQuestionFindAllResponseToJSONTyped(json, false);
}

export function CommonResponseScaleQuestionFindAllResponseToJSONTyped(value?: CommonResponseScaleQuestionFindAllResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'code': value['code'],
        'message': value['message'],
        'data': ScaleQuestionFindAllResponseToJSON(value['data']),
    };
}