// src/generated-api/models/ScaleQuestionDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CareDiary API
 * 돌봄일기 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 척도 질문 상세 정보
 * @export
 * @interface ScaleQuestionDto
 */
export interface ScaleQuestionDto {
    /**
     * 척도 질문 ID
     * @type {number}
     * @memberof ScaleQuestionDto
     */
    scaleQuestionId: number;
    /**
     * 질문 번호
     * @type {number}
     * @memberof ScaleQuestionDto
     */
    questionNumber: number;
    /**
     * 질문 내용
     * @type {string}
     * @memberof ScaleQuestionDto
     */
    content: string;
    /**
     * 척도 카테고리 (ANXIETY, DEPRESSION, ANGER)
     * @type {string}
     * @memberof ScaleQuestionDto
     */
    scaleCategory: ScaleQuestionDtoScaleCategoryEnum;
    /**
     * 선택지 개수
     * @type {number}
     * @memberof ScaleQuestionDto
     */
    optionCount: number;
    /**
     * 선택지 목록
     * @type {Array<string>}
     * @memberof ScaleQuestionDto
     */
    options: Array<string>;
}


/**
 * @export
 */
export const ScaleQuestionDtoScaleCategoryEnum = {
    Anxiety: 'ANXIETY',
    Depression: 'DEPRESSION',
    Anger: 'ANGER'
} as const;
export type ScaleQuestionDtoScaleCategoryEnum = typeof ScaleQuestionDtoScaleCategoryEnum[keyof typeof ScaleQuestionDtoScaleCategoryEnum];


/**
 * Check if a given object implements the ScaleQuestionDto interface.
 */
export function instanceOfScaleQuestionDto(value: object): value is ScaleQuestionDto {
    if (!('scaleQuestionId' in value) || value['scaleQuestionId'] === undefined) return false;
    if (!('questionNumber' in value) || value['questionNumber'] === undefined) return false;
    if (!('content' in value) || value['content'] === undefined) return false;
    if (!('scaleCategory' in value) || value['scaleCategory'] === undefined) return false;
    if (!('optionCount' in value) || value['optionCount'] === undefined) return false;
    if (!('options' in value) || value['options'] === undefined) return false;
    return true;
}

export function ScaleQuestionDtoFromJSON(json: any): ScaleQuestionDto {
    return ScaleQuestionDtoFromJSONTyped(json, false);
}

export function ScaleQuestionDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ScaleQuestionDto {
    if (json == null) {
        return json;
    }
    return {
        
        'scaleQuestionId': json['scaleQuestionId'],
        'questionNumber': json['questionNumber'],
        'content': json['content'],
        'scaleCategory': json['scaleCategory'],
        'optionCount': json['optionCount'],
        'options': json['options'],
    };
}

export function ScaleQuestionDtoToJSON(json: any): ScaleQuestionDto {
    return ScaleQuestionDtoToJSONTyped(json, false);
}

export function ScaleQuestionDtoToJSONTyped(value?: ScaleQuestionDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'scaleQuestionId': value['scaleQuestionId'],
        'questionNumber': value['questionNumber'],
        'content': value['content'],
        'scaleCategory': value['scaleCategory'],
        'optionCount': value['optionCount'],
        'options': value['options'],
    };
}