// src/generated-api/models/ScaleQuestionUserAnswerItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CareDiary API
 * 돌봄일기 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 척도 질문 응답 항목
 * @export
 * @interface ScaleQuestionUserAnswerItem
 */
export interface ScaleQuestionUserAnswerItem {
    /**
     * 척도 질문 ID
     * @type {number}
     * @memberof ScaleQuestionUserAnswerItem
     */
    scaleQuestionId: number;
    /**
     * 응답 값 (1-5)
     * @type {number}
     * @memberof ScaleQuestionUserAnswerItem
     */
    answer: number;
}

/**
 * Check if a given object implements the ScaleQuestionUserAnswerItem interface.
 */
export function instanceOfScaleQuestionUserAnswerItem(value: object): value is ScaleQuestionUserAnswerItem {
    if (!('scaleQuestionId' in value) || value['scaleQuestionId'] === undefined) return false;
    if (!('answer' in value) || value['answer'] === undefined) return false;
    return true;
}

export function ScaleQuestionUserAnswerItemFromJSON(json: any): ScaleQuestionUserAnswerItem {
    return ScaleQuestionUserAnswerItemFromJSONTyped(json, false);
}

export function ScaleQuestionUserAnswerItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): ScaleQuestionUserAnswerItem {
    if (json == null) {
        return json;
    }
    return {
        
        'scaleQuestionId': json['scaleQuestionId'],
        'answer': json['answer'],
    };
}

export function ScaleQuestionUserAnswerItemToJSON(json: any): ScaleQuestionUserAnswerItem {
    return ScaleQuestionUserAnswerItemToJSONTyped(json, false);
}

export function ScaleQuestionUserAnswerItemToJSONTyped(value?: ScaleQuestionUserAnswerItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'scaleQuestionId': value['scaleQuestionId'],
        'answer': value['answer'],
    };
}